requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
bce-python-sdk==0.9.60
bcrypt==3.2.2
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.4
cssselect==1.4.0
cssutils==2.11.1
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
easyocr==1.7.2
ecdsa==0.19.1
einops==0.8.2
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
future==1.0.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx>=0.23.0
httpx-sse==0.4.3
huggingface_hub==1.3.5
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.7.3
lazy_loader==0.4
limits==5.8.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
modelscope==1.34.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
numpy==2.2.6
openai==2.17.0
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.7
packaging==25.0
paddleocr==3.4.0
paddlepaddle==3.2.0
paddlex==3.4.2
pandas==2.3.3
passlib==1.7.4
pdf2image==1.17.0
pillow==12.1.0
premailer==3.10.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.2
pvporcupine==4.0.1
py-cpuinfo==9.0.0
pyasn1==0.6.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
PyGetWindow==0.0.9
PyJWT==2.11.0
PyMsgBox==2.0.1
PyMuPDF==1.26.7
PyMySQL==1.1.2
pypdf==6.6.2
PyPDF2==3.0.1
pypdfium2==5.4.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
ruamel.yaml==0.19.1
safetensors==0.7.0
schedule==1.2.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
secure==1.0.1
sentence-transformers==5.2.2
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
SpeechRecognition==3.14.5
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
wcwidth==0.6.0
Werkzeug==3.1.5
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
gunicorn>=23.0.0
psycopg2-binary==2.9.9
Flask-Mail==0.10.0
flask-socketio==5.5.1
eventlet>=0.40.3
deep-translator==1.11.4
